video_originality_analyzer/requirements.txt
fastapi
uvicorn
pydantic
yt-dlp
opencv-python-headless
numpy
Pillow
ImageHash
